# core/generated/swagger-ui/schema-registry-api-spec.yaml
openapi: 3.0.1
info:
  title: Confluent Schema Registry
  version: v1
paths:
...
  /schemas:
    get:
      summary: List schemas
      description: Get the schemas matching the specified parameters.
      operationId: getSchemas
      parameters:
      - name: subjectPrefix
        in: query
        description: Filters results by the respective subject prefix
        schema:
          type: string
          default: ""
      - name: deleted
        in: query
        description: Whether to return soft deleted schemas
        schema:
          type: boolean
          default: false
      - name: latestOnly
        in: query
        description: Whether to return latest schema versions only for each matching
          subject
        schema:
          type: boolean
          default: false
      - name: offset
        in: query
        description: Pagination offset for results
        schema:
          type: integer
          format: int32
          default: 0
      - name: limit
        in: query
        description: Pagination size for results. Ignored if negative
        schema:
          type: integer
          format: int32
          default: -1
      responses:
        "200":
          description: The schemas matching the specified parameters
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: array
                items:
                  type: string
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: array
                items:
                  type: string
            application/json; qs=0.5:
              schema:
                type: array
                items:
                  type: string
        "500":
          description: |
            Error code 50001 -- Error in the backend data store
  /schemas/ids/{id}:
    get:
      summary: Get schema string by ID
      description: Retrieves the schema string identified by the input ID.
      operationId: getSchema
      parameters:
      - name: id
        in: path
        description: Globally unique identifier of the schema
        required: true
        schema:
          type: integer
          format: int32
      - name: subject
        in: query
        description: Name of the subject
        schema:
          type: string
      - name: format
        in: query
        description: "Desired output format, dependent on schema type"
        schema:
          type: string
          default: ""
      - name: fetchMaxId
        in: query
        description: Whether to fetch the maximum schema identifier that exists
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: The schema string
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                $ref: '#/components/schemas/SchemaString'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                $ref: '#/components/schemas/SchemaString'
            application/json; qs=0.5:
              schema:
                $ref: '#/components/schemas/SchemaString'
        "404":
          description: |
            Error code 40403 -- Schema not found
        "500":
          description: |
            Error code 50001 -- Error in the backend data store
  /schemas/ids/{id}/subjects:
    get:
      summary: List subjects associated to schema ID
      description: Retrieves all the subjects associated with a particular schema
        ID.
      operationId: getSubjects
      parameters:
      - name: id
        in: path
        description: Globally unique identifier of the schema
        required: true
        schema:
          type: integer
          format: int32
      - name: subject
        in: query
        description: Filters results by the respective subject
        schema:
          type: string
      - name: deleted
        in: query
        description: Whether to include subjects where the schema was deleted
        schema:
          type: boolean
      responses:
        "200":
          description: The subjects matching the specified parameters
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: array
                items:
                  type: string
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: array
                items:
                  type: string
            application/json; qs=0.5:
              schema:
                type: array
                items:
                  type: string
        "404":
          description: |
            Error code 40403 -- Schema not found
        "500":
          description: |
            Error code 50001 -- Error in the backend data store
  /schemas/ids/{id}/versions:
    get:
      summary: List subject-versions associated to schema ID
      description: Get all the subject-version pairs associated with the input ID.
      operationId: getVersions
      parameters:
      - name: id
        in: path
        description: Globally unique identifier of the schema
        required: true
        schema:
          type: integer
          format: int32
      - name: subject
        in: query
        description: Filters results by the respective subject
        schema:
          type: string
      - name: deleted
        in: query
        description: Whether to include subject versions where the schema was deleted
        schema:
          type: boolean
      responses:
        "200":
          description: The subject versions matching the specified parameters
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectVersion'
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectVersion'
            application/json; qs=0.5:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SubjectVersion'
        "404":
          description: |
            Error code 40403 -- Schema not found
        "500":
          description: |
            Error code 50001 -- Error in the backend data store
  /schemas/types:
    get:
      summary: List supported schema types
      description: Retrieve the schema types supported by this registry.
      operationId: getSchemaTypes
      responses:
        "200":
          description: The supported schema types
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: array
                items:
                  type: string
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: array
                items:
                  type: string
            application/json; qs=0.5:
              schema:
                type: array
                items:
                  type: string
        "500":
          description: |
            Error code 50001 -- Error in the backend data store
...
  /subjects/{subject}/versions/{version}/schema:
    get:
      summary: Get schema string by version
      description: Retrieves the schema for the specified version of this subject.
        Only the unescaped schema string is returned.
      operationId: getSchemaOnly_2
      parameters:
      - name: subject
        in: path
        description: Name of the subject
        required: true
        schema:
          type: string
      - name: version
        in: path
        description: "Version of the schema to be returned. Valid values for versionId\
          \ are between [1,2^31-1] or the string \"latest\". \"latest\" returns the\
          \ last registered schema under the specified subject. Note that there may\
          \ be a new latest schema that gets registered right after this request is\
          \ served."
        required: true
        schema:
          type: string
      - name: deleted
        in: query
        description: Whether to include deleted schema
        schema:
          type: boolean
      responses:
        "200":
          description: The schema string
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: string
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: string
            application/json; qs=0.5:
              schema:
                type: string
        "404":
          description: |-
            Error code 40401 -- Subject not found
            Error code 40402 -- Version not found
        "422":
          description: Error code 42202 -- Invalid version
        "500":
          description: Error code 50001 -- Error in the backend data store
...
  /schemas/ids/{id}/schema:
    get:
      summary: Get schema by ID
      description: Retrieves the schema identified by the input ID.
      operationId: getSchemaOnly
      parameters:
      - name: id
        in: path
        description: Globally unique identifier of the schema
        required: true
        schema:
          type: integer
          format: int32
      - name: subject
        in: query
        description: Name of the subject
        schema:
          type: string
      - name: format
        in: query
        description: "Desired output format, dependent on schema type"
        schema:
          type: string
          default: ""
      - name: fetchMaxId
        in: query
        description: Whether to fetch the maximum schema identifier that exists
        schema:
          type: boolean
          default: false
      responses:
        "200":
          description: schema
          content:
            application/vnd.schemaregistry.v1+json:
              schema:
                type: string
            application/vnd.schemaregistry+json; qs=0.9:
              schema:
                type: string
            application/json; qs=0.5:
              schema:
                type: string
        "404":
          description: |
            Error code 40403 -- Schema not found
        "500":
          description: |
            Error code 50001 -- Error in the backend data store
...
components:
  schemas:
...
    SchemaReference:
      type: object
      properties:
        name:
          type: string
          description: Reference name
        subject:
          type: string
          description: Name of the referenced subject
        version:
          type: integer
          description: Version number of the referenced subject
          format: int32
      description: Schema reference
...
    SchemaString:
      type: object
      properties:
        schemaType:
          type: string
          description: Schema type
        schema:
          type: string
          description: Schema string identified by the ID
        references:
          type: array
          description: References to other schemas
          items:
            $ref: '#/components/schemas/SchemaReference'
        maxId:
          type: integer
          description: Maximum ID
          format: int32
      description: Schema definition
...
    SubjectVersion:
      type: object
      properties:
        subject:
          type: string
          description: Name of the subject
        version:
          type: integer
          description: Version number
          format: int32
      description: Subject version pair
...